+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This ChangeLog follows the Keep a ChangeLog guidelines](https://keepachangelog.c
 ### Added
 - default_config.toml is now read at compile time and contains a default config file with some endpoints like dbpedia, wikidata, uniprot, etc. 
 - added list_endpoints to PyRudof
+- Improved visualization of node information using termtree
 
 
 ## v0.1.121